third_party/blink/web_tests/external/wpt/scroll-animations/css/scroll-timeline-cssom.tentative-expected.txt

This is a testharness.js-based test.
Found 107 tests; 13 PASS, 94 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL Empty block assert_equals: expected 1 but got 0
FAIL EOF ends block assert_equals: expected 1 but got 0
FAIL Timeline name can be a <string> assert_equals: expected 1 but got 0
PASS Missing prelude
PASS Missing block
PASS Missing timeline name
PASS Timeline name must be an identifier
PASS Timeline name must match <custom-ident>
PASS Timeline name must match <custom-ident> (caps)
PASS Timeline name must match <custom-ident> (mixed)
PASS Timeline name may not be initial
PASS Timeline name may not be inherit
PASS Timeline name may not be unset
PASS Timeline name may not be revert
PASS Timeline name may not be default
PASS Extra timeline name
FAIL CSSRule.type returns 0 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.name foo assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.name Foo assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.name f___123 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.name a\9 b assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.name "foo" assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.name "none" assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(#foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector( #foo ) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source  selector(#foo)  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source none assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source  none  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(#a\9 b) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source auto assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source #foo assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source element(#foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(#foo more) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(html) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(:before) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(*) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(.a) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.source selector(.a, .b) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation auto assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation block assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation inline assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation horizontal assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation vertical assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation   vertical   assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation foo assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation 10px assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.orientation red assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start auto assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end auto assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start  auto  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end  auto  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start 10px assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end 10px assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start  10px  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end  10px  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start 10em assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end 10em assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start 10% assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end 10% assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start calc(1px + 1%) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end calc(1px + 1%) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start red assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end red assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start #fff assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end #fff assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start unset assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end unset assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start unknown(#foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end unknown(#foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start start assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end start assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start end assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end end assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start 3 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end 3 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start selector(#foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end selector(#foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start  selector(#foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end  selector(#foo) assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start selector(#foo) start assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end selector(#foo) start assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start selector(#foo) start 3 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end selector(#foo) start 3 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start selector(#foo) 3 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end selector(#foo) 3 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start selector(#foo) 3.5 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end selector(#foo) 3.5 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start selector(#foo) end assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end selector(#foo) end assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start selector(#foo) end 3 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end selector(#foo) end 3 assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.start selector(#foo) 3 end assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.end selector(#foo) 3 end assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange auto assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange  auto  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange 1s assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange  1s  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange 1000ms assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange  assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange red assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange #fff assert_equals: expected 1 but got 0
FAIL CSSScrollTimelineRule.timeRange unset assert_equals: expected 1 but got 0
Harness: the test ran to completion.
